refactor(suggestions): group src into engine / index modules
Reorganize funput-suggestions/src from 10 flat files into concern-based
module groups, mirroring the funput-ffi / funput-jni layout. Pure internal
reorganization — no public API or behavior change.

- engine/: SuggestionEngine facade (mod.rs) + config + the learn / query /
  durability behaviors (durability.rs was storage.rs, renamed to avoid clashing
  with persistence/)
- index/: the in-memory search index (trie, ranking, normalize)
- types.rs stays at the crate root as the shared vocabulary, keeping every
  crate::types:: path stable
- persistence/ and tests/ unchanged

Every file <= 150 LOC and <= 5 files per directory. Public API
(SuggestionEngine, SuggestionConfig, LearnOutcome, SuggestionSet,
SuggestionStats) is preserved, so funput-ffi, funput-jni, and the
examples/benches/integration tests are untouched.
