+++ crates/funput-suggestions/src/engine/config.rs
