+++ crates/funput-suggestions/src/index/normalize.rs
