+++ crates/funput-suggestions/src/lib.rs
@@ -2,20 +2,21 @@
 //!
 //! This crate is deliberately independent from the Vietnamese composition engine.
 //! Platforms own it on a background worker; lookup never performs I/O.
+//!
+//! # Layout
+//!
+//! - `engine/` — the [`SuggestionEngine`] facade, its [`SuggestionConfig`], and the
+//!   learn / query / durability behaviours.
+//! - `index/` — the in-memory search index (prefix trie, ranking, key normalization).
+//! - `persistence/` — the crash-safe on-disk snapshot + journal store.
+//! - `types` — the shared vocabulary (`WordRecord` and the public result types).
 
-mod config;
 mod engine;
-mod learning;
-mod normalize;
+mod index;
 mod persistence;
-mod query;
-mod ranking;
-mod storage;
-mod trie;
 mod types;
 
-pub use config::SuggestionConfig;
-pub use engine::SuggestionEngine;
+pub use engine::{SuggestionConfig, SuggestionEngine};
 pub use types::{LearnOutcome, SuggestionSet, SuggestionStats};
 
 #[cfg(test)]
